comments; types; fields are hidden in cfg file

+++ exonum/src/api/backends/actix.rs
@@ -297,10 +297,11 @@ pub struct SystemRuntimeConfig {
     pub api_runtimes: Vec<ApiRuntimeConfig>,
     /// API aggregator.
     pub api_aggregator: ApiAggregator,
-    /// The interval in milliseconds between attempts of restarting HTTP-server in case the server failed to restart
+    /// The interval in milliseconds between attempts of restarting HTTP-server in case
+    /// the server failed to restart
     pub service_retry_interval: u64,
     /// The attempts counts of restarting HTTP-server in case the server failed to restart
-    pub service_retry_attempt: u64,
+    pub service_retry_attempt: u16,
 }
 
 /// Actix system runtime handle.
